PaddleAudio/examples/speaker/requirements.txt
paddlepaddle-gpu==2.1.1
PyYAML==5.4.1
